requirements.txt
python-dotenv
pathlib
keyring
asyncio
httpx[http2]
pyotp
pydantic